Refactoring code base, splitting up components and pages into their own repositories.
