+++ src/app/components/footer/footer.component.html
@@ -1,0 +1,30 @@
+<footer class="footer">
+  <div class="container">
+    <div class="footer-content">
+      <div class="footer-section">
+        <h3>Sorella Home Solutions</h3>
+        <p>Transforming homes with professional craftsmanship and attention to detail.</p>
+      </div>
+      <div class="footer-section">
+        <h4>Services</h4>
+        <ul>
+          <li><a href="#services">Kitchen Remodeling</a></li>
+          <li><a href="#services">Bathroom Renovation</a></li>
+          <li><a href="#services">Interior Design</a></li>
+          <li><a href="#services">Home Maintenance</a></li>
+        </ul>
+      </div>
+      <div class="footer-section">
+        <h4>Contact</h4>
+        <ul>
+          <li>Phone: (555) 123-4567</li>
+          <li>Email: info@sorellahomesolutions.com</li>
+          <li>123 Main Street, Your City, State 12345</li>
+        </ul>
+      </div>
+    </div>
+    <div class="footer-bottom">
+      <p>&copy; 2024 Sorella Home Solutions. All rights reserved.</p>
+    </div>
+  </div>
+</footer>
